+++ include/commands.h
@@ -15,6 +15,7 @@ typedef enum {
     COMMAND_EXPORT,
     COMMAND_EDIT,
     COMMAND_PUSH,
+    COMMAND_GENERATE,
 } Command;
 
 void execute_command(Command command, Options options);
@@ -29,6 +30,7 @@ void command_export(Options options);
 void command_get(Options options);
 void command_edit(Options options);
 void command_push(Options options);
+void command_generate(Options options);
 char* command_to_string(Command command);
 
 #endif // COMMANDS_H